+++ src/components/settings/Settings.vue
@@ -1,9 +1,11 @@
 <script>
+    import { toRaw} from 'vue';
     import { LANGUAGES } from '../../editor/languages.js'
 
     import KeyboardHotkey from "./KeyboardHotkey.vue"
     import TabListItem from "./TabListItem.vue"
     import TabContent from "./TabContent.vue"
+    import KeyboardBindings from './KeyboardBindings.vue'
 
     const defaultFontFamily = window.heynote.defaultFontFamily
     const defaultFontSize = window.heynote.defaultFontSize
@@ -20,15 +22,18 @@
             KeyboardHotkey,
             TabListItem,
             TabContent,
+            KeyboardBindings,
         },
 
         data() {
+            //console.log("settings:", this.initialSettings)
             return {
                 keymaps: [
                     { name: "Default", value: "default" },
                     { name: "Emacs", value: "emacs" },
                 ],
                 keymap: this.initialSettings.keymap,
+                keyBindings: this.initialSettings.keyBindings,
                 metaKey: this.initialSettings.emacsMetaKey,
                 isMac: window.heynote.platform.isMac,
                 showLineNumberGutter: this.initialSettings.showLineNumberGutter,
@@ -63,28 +68,36 @@
                 defaultFontSize: defaultFontSize,
                 appVersion: "",
                 theme: this.themeSetting,
+
+                // tracks if the add key binding dialog is visible (so that we can set inert on the save button)
+                addKeyBindingDialogVisible: false,
             }
         },
 
         async mounted() {
+            window.addEventListener("keydown", this.onKeyDown);
+
+            this.appVersion = await window.heynote.getVersion()
+
             if (window.queryLocalFonts !== undefined) {
                 let localFonts = [... new Set((await window.queryLocalFonts()).map(f => f.family))].filter(f => f !== "Hack")
                 localFonts = [...new Set(localFonts)].map(f => [f, f])
                 this.systemFonts = [[defaultFontFamily, defaultFontFamily + " (default)"], ...localFonts]
             }
-
-            window.addEventListener("keydown", this.onKeyDown);
-            this.$refs.keymapSelector.focus()
-
-            this.appVersion = await window.heynote.getVersion()
         },
         beforeUnmount() {
             window.removeEventListener("keydown", this.onKeyDown);
         },
 
+        watch: {
+            keyBindings(newKeyBindings) {
+                this.updateSettings()
+            }
+        },
+
         methods: {
             onKeyDown(event) {
-                if (event.key === "Escape") {
+                if (event.key === "Escape" && !this.addKeyBindingDialogVisible) {
                     this.$emit("closeSettings")
                 }
             },
@@ -94,6 +107,7 @@
                     showLineNumberGutter: this.showLineNumberGutter,
                     showFoldGutter: this.showFoldGutter,
                     keymap: this.keymap,
+                    keyBindings: this.keyBindings.map((kb) => toRaw(kb)),
                     emacsMetaKey: window.heynote.platform.isMac ? this.metaKey : "alt",
                     allowBetaVersions: this.allowBetaVersions,
                     enableGlobalHotkey: this.enableGlobalHotkey,
@@ -162,6 +176,12 @@
                             @click="activeTab = 'appearance'"
                         />
                         <TabListItem 
+                            name="Key Bindings" 
+                            tab="keyboard-bindings" 
+                            :activeTab="activeTab" 
+                            @click="activeTab = 'keyboard-bindings'"
+                        />
+                        <TabListItem 
                             :name="isWebApp ? 'Version' : 'Updates'" 
                             tab="updates" 
                             :activeTab="activeTab" 
@@ -171,23 +191,6 @@
                 </nav>
                 <div class="settings-content">
                     <TabContent tab="general" :activeTab="activeTab">
-                        <div class="row">
-                            <div class="entry">
-                                <h2>Keymap</h2>
-                                <select ref="keymapSelector" v-model="keymap" @change="updateSettings" class="keymap">
-                                    <template v-for="km in keymaps" :key="km.value">
-                                        <option :selected="km.value === keymap" :value="km.value">{{ km.name }}</option>
-                                    </template>
-                                </select>
-                            </div>
-                            <div class="entry" v-if="keymap === 'emacs' && isMac">
-                                <h2>Meta Key</h2>
-                                <select v-model="metaKey" @change="updateSettings" class="metaKey">
-                                    <option :selected="metaKey === 'meta'" value="meta">Command</option>
-                                    <option :selected="metaKey === 'alt'" value="alt">Option</option>
-                                </select>
-                            </div>
-                        </div>
                         <div class="row" v-if="!isWebApp">
                             <div class="entry">
                                 <h2>Global Keyboard Shortcut</h2>
@@ -244,14 +247,14 @@
                         </div>
                         <div class="row" v-if="!isWebApp">
                             <div class="entry buffer-location">
-                                <h2>Buffer File Path</h2>
+                                <h2>Buffer Files Path</h2>
                                 <label class="keyboard-shortcut-label">
                                     <input 
                                         type="checkbox" 
                                         v-model="customBufferLocation" 
                                         @change="onCustomBufferLocationChange"
                                     />
-                                    Use custom buffer file location
+                                    Use custom location for the buffer files
                                 </label>
                                 <div class="file-path">
                                     <button
@@ -368,6 +371,31 @@
                             </div>
                         </div>
                     </TabContent>
+
+                    <TabContent tab="keyboard-bindings" :activeTab="activeTab">
+                        <div class="row">
+                            <div class="entry">
+                                <h2>Keymap</h2>
+                                <select v-model="keymap" @change="updateSettings" class="keymap">
+                                    <template v-for="km in keymaps" :key="km.value">
+                                        <option :selected="km.value === keymap" :value="km.value">{{ km.name }}</option>
+                                    </template>
+                                </select>
+                            </div>
+                            <div class="entry" v-if="keymap === 'emacs' && isMac">
+                                <h2>Meta Key</h2>
+                                <select v-model="metaKey" @change="updateSettings" class="metaKey">
+                                    <option :selected="metaKey === 'meta'" value="meta">Command</option>
+                                    <option :selected="metaKey === 'alt'" value="alt">Option</option>
+                                </select>
+                            </div>
+                        </div>
+                        <KeyboardBindings 
+                            :userKeys="keyBindings ? keyBindings : {}"
+                            v-model="keyBindings"
+                            @addKeyBindingDialogVisible="addKeyBindingDialogVisible = $event"
+                        />
+                    </TabContent>
                     
                     <TabContent tab="updates" :activeTab="activeTab">
                         <div class="row">
@@ -407,7 +435,7 @@
                 </div>
             </div>
             
-            <div class="bottom-bar">
+            <div class="bottom-bar" :inert="addKeyBindingDialogVisible">
                 <button 
                     @click="$emit('closeSettings')"
                     class="close"
@@ -436,14 +464,16 @@
             background: rgba(0, 0, 0, 0.5)
         
         .dialog
+            --dialog-height: 600px
+            --bottom-bar-height: 48px
             box-sizing: border-box
             z-index: 2
             position: absolute
             left: 50%
             top: 50%
             transform: translate(-50%, -50%)
-            width: 700px
-            height: 560px
+            width: 820px
+            height: var(--dialog-height)
             max-width: 100%
             max-height: 100%
             display: flex
@@ -463,6 +493,7 @@
             .dialog-content
                 flex-grow: 1
                 display: flex
+                height: calc(var(--dialog-height) - var(--bottom-bar-height))
                 .sidebar
                     box-sizing: border-box
                     width: 140px
@@ -480,6 +511,7 @@
                     flex-grow: 1
                     padding: 40px
                     overflow-y: auto
+                    position: relative
                     select
                         height: 22px
                         margin: 4px 0
@@ -536,6 +568,8 @@
                                         background: #222
                                         color: #aaa
             .bottom-bar
+                box-sizing: border-box
+                height: var(--bottom-bar-height)
                 border-radius: 0 0 5px 5px
                 background: #eee
                 text-align: right
@@ -544,4 +578,5 @@
                     background: #222
                 .close
                     height: 28px
+                    cursor: pointer
 </style>