+++ src/editor/annotation.js
@@ -8,3 +8,10 @@ export const ADD_NEW_BLOCK = "heynote-add-new-block"
 export const DELETE_BLOCK = "heynote-delete-block"
 export const CURSOR_CHANGE = "heynote-cursor-change"
 export const APPEND_BLOCK = "heynote-append-block"
+export const SET_FONT = "heynote-set-font"
+
+
+// This function checks if any of the transactions has the given Heynote annotation
+export function transactionsHasAnnotation(transactions, annotation) {
+    return transactions.some(tr => tr.annotation(heynoteEvent) === annotation)
+}