+++ src/main.js
@@ -1,7 +1,9 @@
 import './css/application.sass'
+import '../assets/font/open-sans/open-sans.css'
 
 import { createApp } from 'vue'
 import { createPinia } from 'pinia'
+import PrimeVue from 'primevue/config';
 
 import App from './components/App.vue'
 import { loadCurrencies } from './currency'
@@ -13,6 +15,7 @@ const pinia = createPinia()
 const app = createApp(App)
 
 app.use(pinia)
+app.use(PrimeVue)
 app.mount('#app').$nextTick(() => {
     // hide loading screen
     postMessage({ payload: 'removeLoading' }, '*')